+++ server/src/repositories/checks/IChecksRepository.ts
@@ -1,57 +1,5 @@
-import type { Check, MonitorType } from "@/types/index.js";
+import type { Check, ChecksQueryResult, ChecksSummary, MonitorType, PageSpeedChecksResult, HardwareChecksResult, UptimeChecksResult } from "@/types/index.js";
 import type { LatestChecksMap } from "@/repositories/checks/MongoChecksRepistory.js";
-
-export interface PageSpeedChecksResult {
-	monitorType: "pagespeed";
-	checks: Check[];
-}
-
-export interface HardwareChecksResult {
-	monitorType: "hardware";
-	aggregateData: {
-		latestCheck: Check | null;
-		totalChecks: number;
-	};
-	upChecks: {
-		totalChecks: number;
-	};
-	checks: Array<{
-		_id: string;
-		avgCpuUsage: number;
-		avgMemoryUsage: number;
-		avgTemperature: number[];
-		disks: Array<{
-			name: string;
-			readSpeed: number;
-			writeSpeed: number;
-			totalBytes: number;
-			freeBytes: number;
-			usagePercent: number;
-		}>;
-		net: Array<{
-			name: string;
-			bytesSentPerSecond: number;
-			deltaBytesRecv: number;
-			deltaPacketsSent: number;
-			deltaPacketsRecv: number;
-			deltaErrIn: number;
-			deltaErrOut: number;
-			deltaDropIn: number;
-			deltaDropOut: number;
-			deltaFifoIn: number;
-			deltaFifoOut: number;
-		}>;
-	}>;
-}
-
-export interface UptimeChecksResult {
-	monitorType: Exclude<MonitorType, "hardware" | "pagespeed">;
-	groupedChecks: Array<{ _id: string; avgResponseTime: number; totalChecks: number }>;
-	groupedUpChecks: Array<{ _id: string; totalChecks: number; avgResponseTime: number }>;
-	groupedDownChecks: Array<{ _id: string; totalChecks: number; avgResponseTime: number }>;
-	uptimePercentage: number;
-	avgResponseTime: number;
-}
 
 export interface IChecksRepository {
 	// create
@@ -67,8 +15,8 @@ export interface IChecksRepository {
 		page: number,
 		rowsPerPage: number,
 		status: boolean | undefined
-	): Promise<any>;
-	findByTeamId(sortOrder: string, dateRange: string, filter: string, page: number, rowsPerPage: number, teamId: string): Promise<any>;
+	): Promise<ChecksQueryResult>;
+	findByTeamId(sortOrder: string, dateRange: string, filter: string, page: number, rowsPerPage: number, teamId: string): Promise<ChecksQueryResult>;
 	findLatestByMonitorIds(monitorIds: string[], options?: { limitPerMonitor?: number }): Promise<LatestChecksMap>;
 	findByDateRangeAndMonitorId(
 		monitorId: string,
@@ -77,7 +25,7 @@ export interface IChecksRepository {
 		dateString: string,
 		options?: { type?: MonitorType }
 	): Promise<UptimeChecksResult | HardwareChecksResult | PageSpeedChecksResult>;
-	findSummaryByTeamId(teamId: string): Promise<any>;
+	findSummaryByTeamId(teamId: string): Promise<ChecksSummary>;
 	// update
 	//delete
 	deleteByMonitorId(monitorId: string): Promise<number>;
